fixed issues with adding file to target
